cauldron: chroots are tracked with list of users
Chroots are now treated as single objects that can have multiple
dependees, or users. When cauldron_chroot_init is called on a chroot
that is already initialized, the user count is increased, rather than
producing an error. When a subsequent cauldron_chroot_done is
encountered, the user count is checked, and the chroot is cleaned up
only if the final user is the one that made the call. This also enables
a sort of context-switching for cauldron_chroot calls.