cauldron: fix the logic in lib.chroot
Although lib.chroot now tracks the number of users/dependees of each
chroot, the logic for detecting if the necessary chroot was already
loaded, and what to do if it needed to be loaded, was still based on the
old functionality of only having a single chroot with a single
user/dependee at a time. Additionally, it's wasteful to perform the load
operations on a chroot that is already loaded, so it should only be done
for the first user. Thus, this commit also optimizes the algorithms a
bit.