cauldron: test if chroot successfully loaded
The previous commit more properly initializes/cleans up chroots based on
the users/dependees. That only leaves the chroot detection code to be
updated. This commit introduces a new function specifically for testing
whether there is a currently loaded chroot, and can also test whether it
matches a given path to ensure that a desired chroot is the current one.
This commit also brings the rest of the libraries that depend on chroot
functionality in sync with the new code, including updating the chroot
error code definitions.