implement true library function definition tests
Functions were tested for their definition by simply testing if their
names were null (-z). This was incorrect, however, as bare-words tested
for null will always return true, making the test a no-op. The tests
have been reimplemented using bash's type mechanism, which will return
the type of the variable/name, which for a defined function would be the
word "function", but for something undefined it returns an empty string.
It also sets the return value ($?) to true or false depending on whether
the variable/name is defined.