use liberror_check for better error checking
Rather than simply returning errors up to the calling functions, make
use of liberror_check. The appearance is not only cleaner, it also
outputs error messages that demonstrate what went wrong and where,
rather than just bailing out and leaving it up to the user to track the
error down.

There are still more points in the codebase to evaluate for better
testing and error handling. There will be further use of the liberror
functions in later commits.