cauldron: config: correct initramfs creation cmd
The initramfs creation command failed to function correctly due to the
way it was parsed. Instead, pass it through an eval to ensure that each
command is properly run, as well as piped to the subsequent commands.

Additionally, add the --quiet flag to cpio to prevent spamming output.
End-users can modify this variable to remove the --quiet flag if they
want to debug the initramfs creation.