errorcodes: change file/dir NOT_FOUND -> ACCESS
Change error code naming and usage semantics to file/dir access. Some of the
errors are hard to specifically detect for whether files are actually
non-existent or whether they can't be accessed due to permissions problems. Now
that both original error codes from called programs as well as more complete
error messages (including source file and line numbers) are percolated up in
the error reporting system, we can afford to not be specific about the cause of
the error (non-existence vs. insufficient access rights) and allow the end-user
to discern that cause.