cauldron: config/arch: add sorcery arch settings
Different parts of the system use different names for different architectures.
For example, the linux kernel does not refer to specific x86 implementations
(and does not understand them, rejecting them), but the compilers (gcc, llvm)
do, so they need to be given i486 (or similar) instead of x86. Since these are
all arch-specific settings, add new variables in the arch configs to hold the
different values for different major architectures.