enchantment: rename module text files
Instead of redundantly naming module files after the name of the module itself,
give them more meaningful names that are not based on the module name. If the
module is ranamed, it will not necessitate changes to the filenames of files
belonging to that module, and it should be more readily apparent what the
purpose of each file is, making modifications and direct manipulations by end
users easier.