cauldron: add function to initialize fs dirs
System filesystem initialization requires more than just adding binaries to
requisite paths. Various directories need to be present with their associated
permissions, and various files (especially device files) are also needed. This
commit adds a function to libcauldron that ensures a given chroot path is
populated with the minimum required directories, sets the permissions on those
directories, as well as creates absolutely necessary device nodes in the dev/
tree. It also runs ldconfig within the chroot to ensure that library searching
has the correct path and that libraries (especially the system libc) will be
usable.

Currently this function does not rely on any spells, such as smgl-fhs, for
bootstrapping the target chroot path. Such a dependency may be desired in the
future, after testing to determine if the spell can provide the same
requirements as cauldron without too many additions.