cauldron: lib.sorcerous: major overhaul
Refactored and reimplemented almost all of the sorcery interface functionality.
The interface is now cleaner, and allows for various sorcery operations to be
performed on different target chroot directories. Additionally, specific
functionality was implemented for configuring sorcery with different settings
depending on the purpose of the given chroot (for example, a sys chroot should
have mostly unmodified sorcery defaults, except for having ARCHITECTURE set
appropriately).