ward: cleaner options and sorting
For checking options to functions, rather than including the leading hyphen in
each case statement, strip the leading hyphen and only test for the option
value itself.

For sorting and removing duplicates of variables, echo can be problematic, but
printf is not (and it is also more portable). Alter the functions to use
printf, removing the need for awk.