cauldron: lib.toolchain: refactor + reimplement
The toolchain code was both incomplete and a mess as it was. This implements
missing functionality required to actually build a toolchain from scratch,
including acquiring the requisite sources for the kernel (required for header
files at least, needed to build the C standard library) and compilers. Unneeded
functionality was also stripped out.