cauldron: lib.init: path in local variable
Push the sorcery path to operate on into a local variable at the top of the
function so it can be found more easily later if it needs to be modified.

This might be a candidate for an additional global config variable, but
currently since sorcery has its paths hardcoded this isn't possible, so it's
left as a local variable for now.

Also add an additional check to make sure the target initfs directory exists
before operating on/within it.