cauldron: lib.init: ensure sh symlink at boot
When a system boots from the final iso product, the kernel will load the initfs
and try to start the init process therein. The init, which in cauldron is a
shell script, cannot run if there is no shell that can be started. The
bootloader could be changed (or the shell script's shabang) to indicate the
sysutil program in order to spawn the shell, but it's better to ensure a
symlink is present and use a more standard /bin/sh instead.

This could additionally foster run-time testing, as a simple shell script could
be executed by the initfs shell (the symlink to the sysutil) to ensure it works
before packaging it into the final iso.