cauldron: complete libcauldron implementation
The component libraries for more specific functionality are implemented, but
the core that puts it all together, both through convenience functions to the
component libraries as well as miscellaneous functionality of its own, was
largely incomplete. This is a large commit, but testing individual functions
was difficult without a produced final iso, so the implemented functions are
being committed together.