cauldron: add misc etc config/data files
In order to produce a working iso, the iso needs various files to populate its
etc/ tree. Cauldron also needs to configure the toolchain with specific
settings, particularly for sorcery, and the initrd/initramfs that will boot the
iso also needs various files (including its init script). This commit adds
these necessary files.

Many of the files are in locations that may change in a future update. In other
words, the locations of these files (or even the fact that the content is
provided in a file instead of a global config shell variable) should not be
considered stable. Additionally, some content may be found unnecessary or
incorrect/incomplete and hence later removed or altered.