cauldron: (re-)implement the end-user program
(Re)-implement the end-user cauldron program/command. The previous
implementation was really just a stub with almost no functionality, with a
skeleton that mirrored that of sorcery's. As cauldron has evolved away from
sorcery, most of the skeleton framework that no longer applies was stripped
out, and functions were filled in to interact with all the parts of libcauldron
and the other component libraries to produce a final iso.

A lot of functionality is yet to be implemented, with stub functions especially
for update_* features. These funstions will need to be implemented in the
future, and the program could use a good refactoring on the whole.