cauldron: initial commit of cauldron man page
This is the initial commit of the cauldron man page. Only some of the existing
features/functions are currently documented.

This will get fleshed out and updated later as time permits, but it's better to
commit the documentation that exists than to have no documentation at all for
all the new functionality.