enchantment: rename shell enchanter binary
Change the name of the enchanter binary from enchantment to enchantment-shell.

All enchanters should be named for themselves, rather than taking the generic
"enchantment" name. The generic name will appear as a symlink to the main
enchanter in the final iso. This is to allow for multiple different enchantment
interfaces in a single iso.