cauldron: add output error code
Add an error code for when output cannot be written to the display device/file.
When the display is a terminal, it is very unlikely that an error can occur.
However, if the user (re)configures cauldron's output to direct to a file, it
can be possible that the file cannot be written to due to insufficient
permissions or an overfull filesystem, for example.