enchantment: enchance config/display.conf
Add additional variables to display.conf for controlling both the level of
output displayed (ENCHANT_VERBOSE) and where the output will be written to
(NULL, STDOUT, OR STDERR). The output location variables are actually
conditionally defined in various enchantment libraries, but this gives the user
a central controlling point without having to read the various library sources
separately.