define config vars using := param substitution
Switch from the FOO=${FOO:-bar} style in sorcery to using the more compact and
less error-prone (at least for typos) : ${FOO:=bar} form. The leading ':' is a
necessary no-op that prevents the shell from trying to interpret the variable
as a command to run rather than a mere value assignment.