enchantment: add version variable and command
Add a configuration variable to set the versions of the scripts to be included
in the final iso, as well as the version of the final iso itself.

If the iso process is put into full production such that a new iso is
automatically produced every time there is a version bump in the grimoire, it
will be possible to configure the iso version to automatically match the
version of the grimoire included in the iso.