enchantment: update and add modules
Renew the instructions for various modules, as well as add some new ones.

Due to the newer, more dynamic nature of iso creation, many of the module
instructions need to be revisited, as they provide only static text with
preconceived notions of what will be available in the iso runtime.