cauldron: lib.chroot: refactor done + cleanup
Refactor the shared code in cauldron_chroot_done and cauldron_chroot_cleanup
into a separate, new function cauldron_chroot_close. The usage of
cauldron_chroot_close is potentially dangerous (whereas cleanup is guaranteed
dangerous), so it should be used carefully.

Additionally, rather than hardcoding the mountpoints or relying on a recursive
umount, which may not be available on the host system, get the list of mount
points to umount from the list of what's actually mounted in that chroot, but
sorted in reverse order (to guarantee that nested subdirs are unmounted before
their parents).